carga-pendencia-backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
openpyxl==3.1.2
numpy==1.24.3
pandas==2.0.3
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1
pika==1.3.2
mysql-connector-python==8.2.0
ocrmypdf==14.0.0
img2pdf==0.5.0
pillow==10.2.0 